Create m.file
In this code, the input image ("grape.jpg") is read and converted to grayscale if necessary. Then, the image is transformed into the frequency domain, and a filter corresponding to the grape points in the image is applied. Finally, the image is transformed back to the spatial domain and displayed.